Do not overwrite windows sboms at the end

+++ daisy_workflows/export/export_disk.sh
@@ -65,6 +65,8 @@ DESTINATION=$(curl -f -H Metadata-Flavor:Google ${URL}/destination)
 SBOM_PATH=$(curl -f -H Metadata-Flavor:Google ${URL}/sbom-path)
 # The gcs root for sbom-util. If empty, do not run sbom generation with sbom-util.
 SBOM_UTIL_GCS_ROOT=$(curl -f -H Metadata-Flavor:Google ${URL}/sbom-util-gcs-root)
+# Mostly used for windows workflows, set to true if the sbom is already generated and non-empty.
+SBOM_ALREADY_GENERATED=$(curl -f -H Metadata-Flavor:Google ${URL}/sbom-already-generated)
 
 # This function fetches the sbom-util executable from the gcs bucket.
 function fetch_sbomutil() {
@@ -113,8 +115,10 @@ function runSBOMGeneration() {
 }
 
 # Always create empty sbom file so workflow copying does not fail
-touch image.sbom.json
-gsutil cp image.sbom.json $SBOM_PATH
+if [ $SBOM_ALREADY_GENERATED != "true" ]; then
+  touch image.sbom.json
+  gsutil cp image.sbom.json $SBOM_PATH
+fi
 # If the sbom-util program location is passed in, generate the sbom.
 if [ $SBOM_UTIL_GCS_ROOT != "" ]; then
   runSBOMGeneration
